# docs/v1/swagger.yaml
basePath: /api/web-api-gateway/v1
definitions:
...
  domain.ErrorResponse:
    properties:
      code:
        type: integer
      data: {}
      is_error:
        type: boolean
      message:
        type: string
    type: object
...
  domain.SignIn:
    properties:
      company_id:
        example: 1
        type: integer
      password:
        example: admin
        maxLength: 255
        minLength: 8
        type: string
      username:
        example: admin
        maxLength: 140
        minLength: 5
        type: string
    required:
    - password
    - username
    type: object
  domain.SignUp:
    properties:
      company_id:
        example: 1
        type: integer
      country:
        example: KZ
        type: string
      email:
        example: dmitry@test.com
        maxLength: 140
        minLength: 5
        type: string
      is_active:
        example: true
        type: boolean
      is_approved:
        example: true
        type: boolean
      is_send_system_notification:
        example: false
        type: boolean
      language:
        example: ru
        type: string
      name:
        example: Дмитрий
        maxLength: 140
        type: string
      password:
        example: "12345678"
        maxLength: 255
        minLength: 8
        type: string
      phone:
        example: "+77777777777"
        maxLength: 140
        minLength: 7
        type: string
      position:
        example: test
        type: string
      role:
        example: user
        type: string
      sections:
        example:
        - full_company_access
        - full_access
        items:
          type: string
        type: array
      username:
        example: dmitry
        maxLength: 140
        minLength: 5
        type: string
    required:
    - email
    - name
    - password
    - phone
    - username
    type: object
  domain.SuccessResponse:
    properties:
      data: {}
      is_error:
        type: boolean
      total_count:
        type: integer
    type: object
  domain.TokenResponse:
    properties:
      expires_in:
        type: integer
      refresh_token:
        type: string
      token:
        type: string
    type: object
  domain.TokensRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
host: 91.243.71.100:8080
info:
  contact:
    email: ru.system.ru@gmail.com
    name: ru.system.ru@gmail.com
  description: API gateway
  title: Web api gateway API
  version: "1.0"
paths:
  /auth:
    post:
      consumes:
      - application/json
      description: |-
        Аутентификация пользователя.
        Только у super admin есть возможность авторизоваться под определенной компанией.
      operationId: sign-in
      parameters:
      - description: Необходимо указать данные для аутентификации пользователя.
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.SignIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Sign in
      tags:
      - auth
  /auth/logout:
    get:
      consumes:
      - application/json
      description: Выход из аккаунта пользователя
      operationId: sign-out
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Sign out
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Обновление access токена
      operationId: refresh-tokens
      parameters:
      - description: Необходимо указать текущие refresh token и sections.
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.TokensRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Refresh tokens
      tags:
      - auth
...
  /register:
    post:
      consumes:
      - application/json
      description: |-
        Регистрация нового пользователя.
        Только у super admin есть возможность добавлять пользователя в другие компании.
        Только у super admin есть возможность давать роль admin пользователю
      operationId: sign-up
      parameters:
      - description: Необходимо указать данные для регистрации нового пользователя.
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.SignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Sign up
      tags:
      - auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"